# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API Series",
        "version": "1.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000/api/v1"
        },
        {
            "url": "http://animeStore.test/api/v1"
        }
    ],
    "paths": {
...
        "/orders": {
            "get": {
                "tags": [
                    "Orders"
                ],
                "summary": "Obtener todas las órdenes",
                "description": "Regresa la información perteneciente a\n     *      todas las órdenes hechas por un usuario en particular",
                "operationId": "App\\Http\\Controllers\\OrderController::index",
                "responses": {
                    "200": {
                        "description": "Operacion exitosa"
                    },
                    "401": {
                        "description": "Unauthorized"
                    }
                },
                "security": [
                    {
                        "Bearer Auth": [
                            "write:projects",
                            "read:projects"
                        ]
                    }
                ]
            },
            "post": {
                "tags": [
                    "Orders"
                ],
                "summary": "Crear una orden",
                "description": "Añade la informacion perteneciente a una orden",
                "operationId": "App\\Http\\Controllers\\OrderController::store",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {}
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "OK"
                    },
                    "404": {
                        "description": "El paquete, como order item, no existe"
                    },
                    "401": {
                        "description": "Unauthorized"
                    }
                },
                "security": [
                    {
                        "Bearer Auth": [
                            "write:projects",
                            "read:projects"
                        ]
                    }
                ]
            }
        },
...
        "/packages/{packages}": {
            "get": {
                "tags": [
                    "Packages"
                ],
                "summary": "Obtener un paquete",
                "description": "Regresa la informacion de un paquete",
                "operationId": "App\\Http\\Controllers\\PackageController::show",
                "parameters": [
                    {
                        "name": "packages",
                        "in": "path",
                        "description": "Paquete id",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Operacion exitosa"
                    },
                    "401": {
                        "description": "Unauthorized"
                    },
                    "404": {
                        "description": "Serie no encontrada"
                    }
                },
                "security": [
                    {
                        "Bearer Auth": [
                            "write:projects",
                            "read:projects"
                        ]
                    }
                ]
            },
            "put": {
                "tags": [
                    "Packages"
                ],
                "summary": "Actualizar un paquete",
                "description": "Regresa la informacion de un paquete ya actualizada",
                "operationId": "App\\Http\\Controllers\\PackageController::update",
                "parameters": [
                    {
                        "name": "packages",
                        "in": "path",
                        "description": "Paquete id",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {}
                        }
                    }
                },
                "responses": {
                    "404": {
                        "description": "Paquete a actualizar no encontrada"
                    },
                    "200": {
                        "description": "Operación exitosa"
                    },
                    "401": {
                        "description": "Unauthorized"
                    }
                },
                "security": [
                    {
                        "Bearer Auth": [
                            "write:projects",
                            "read:projects"
                        ]
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Packages"
                ],
                "summary": "Elimina un paquete",
                "description": "Elimina la información de un paquete",
                "operationId": "App\\Http\\Controllers\\PackageController::destroy",
                "parameters": [
                    {
                        "name": "packages",
                        "in": "path",
                        "description": "Paquete id",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Paquete destruido exitosamente"
                    },
                    "404": {
                        "description": "Paquete no encontrado"
                    },
                    "401": {
                        "description": "Unauthorized"
                    }
                },
                "security": [
                    {
                        "Bearer Auth": [
                            "write:projects",
                            "read:projects"
                        ]
                    }
                ]
            }
        },
...
        "/series/{series}": {
            "get": {
                "tags": [
                    "Series"
                ],
                "summary": "Obtener una serie",
                "description": "Regresa la informacion de una serie",
                "operationId": "getSerie",
                "parameters": [
                    {
                        "name": "series",
                        "in": "path",
                        "description": "Serie id",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Operacion exitosa"
                    },
                    "401": {
                        "description": "Unauthorized"
                    },
                    "404": {
                        "description": "Serie no encontrada"
                    }
                },
                "security": [
                    {
                        "Bearer Auth": [
                            "write:projects",
                            "read:projects"
                        ]
                    }
                ]
            },
            "put": {
                "tags": [
                    "Series"
                ],
                "summary": "Actualizar una serie",
                "description": "Regresa la informacion de una serie ya actualizada",
                "operationId": "App\\Http\\Controllers\\SerieController::update",
                "parameters": [
                    {
                        "name": "series",
                        "in": "path",
                        "description": "Serie id",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {}
                        }
                    }
                },
                "responses": {
                    "404": {
                        "description": "Serie a actualizar no encontrada"
                    },
                    "200": {
                        "description": "Operación exitosa"
                    },
                    "401": {
                        "description": "Unauthorized"
                    }
                },
                "security": [
                    {
                        "Bearer Auth": [
                            "write:projects",
                            "read:projects"
                        ]
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Series"
                ],
                "summary": "Elimina una serie",
                "description": "Elimina la información de una serie",
                "operationId": "App\\Http\\Controllers\\SerieController::destroy",
                "parameters": [
                    {
                        "name": "series",
                        "in": "path",
                        "description": "Serie id",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Serie destruida exitosamente"
                    },
                    "404": {
                        "description": "Serie no encontrada"
                    },
                    "401": {
                        "description": "Unauthorized"
                    }
                },
                "security": [
                    {
                        "Bearer Auth": [
                            "write:projects",
                            "read:projects"
                        ]
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "Bearer Auth": {
                "type": "http",
                "description": "Add bearer token to endpoints",
                "name": "Bearer Token",
                "in": "header",
                "scheme": "bearer"
            }
        }
    }
}